+++ docs/CHANGELOG.md
@@ -123,6 +123,8 @@ Version 1.1.0 (Tamarack): Not yet released
     - `Skip cutscene` (`SKIP_CUTSCENE`)
     - `Quick save` (`QUICK_SAVE`)
     - `Quick load` (`QUICK_LOAD`)
+- Add static bomb code option (combination provided by `eckozero`)
+- Add `sp_staticbombcode` console command to toggle static bomb code option
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file